Refactor to Int the credentialId. Solved some warnings. Changed delete button color. Differentiated the usage of the buttons.

+++ app/src/main/java/com/project/passwordmanager/fragments/ModifyDialogFragment.kt
@@ -14,7 +14,7 @@ import com.project.passwordmanager.factories.ModifyDialogViewModelFactory
 import com.project.passwordmanager.model.CredentialDatabase
 import com.project.passwordmanager.viewmodels.ModifyDialogViewModel
 
-class ModifyDialogFragment(private val credentialId : Long): DialogFragment() {
+class ModifyDialogFragment(private val credentialId : Int): DialogFragment() {
     private var _binding: DialogModifyBinding? = null
     private val binding get() = _binding!!
